backend/requirements.txt
numpy
pandas
fastapi
psycopg2-binary
requests
lxml
uvicorn
pydantic
bcrypt
pyjwt
python-dotenv
sqlalchemy
passlib
python-multipart
sqlmodel